# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Travel Plan API
  description: |
    API documentation for the Travel Plan microservices platform.
    
    ## Authentication
    All API endpoints (except `/auth/login` and `/auth/register`) require a valid JWT token.
    Include the token in the `Authorization` header: `Bearer <token>`
    
    ## Error Handling
    All errors return a consistent response format:
    ```json
    {
      "success": false,
      "message": "Error description",
      "data": null,
      "timestamp": "2024-01-01T12:00:00Z"
    }
    ```
  version: 1.0.0
  contact:
    name: Travel Plan Support
    email: support@travelplan.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: Development server (Kong Gateway)
  - url: https://api.travelplan.com
    description: Production server
...
paths:
...
  /auth/register:
    post:
      tags:
        - Authentication
      summary: Register a new user
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/login:
    post:
      tags:
        - Authentication
      summary: Authenticate user
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid credentials
  /auth/refresh:
    post:
      tags:
        - Authentication
      summary: Refresh access token
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
  /auth/logout:
    post:
      tags:
        - Authentication
      summary: Logout user
      operationId: logoutUser
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Logged out successfully

  # User endpoints
...
  /payments:
    get:
      tags:
        - Payments
      summary: Get all payments
      operationId: getAllPayments
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 10
        - name: userId
          in: query
          schema:
            type: integer
            format: int64
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/PaymentStatus'
      responses:
        '200':
          description: List of payments
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentListResponse'

    post:
      tags:
        - Payments
      summary: Create a new payment
      operationId: createPayment
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePaymentRequest'
      responses:
        '201':
          description: Payment created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentResponse'
  /payments/{id}/process:
    post:
      tags:
        - Payments
      summary: Process a payment
      operationId: processPayment
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Payment processed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentResponse'
  /payments/{id}/refund:
    post:
      tags:
        - Payments
      summary: Refund a payment
      operationId: refundPayment
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                amount:
                  type: number
                  description: Partial refund amount (optional)
      responses:
        '200':
          description: Payment refunded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentResponse'

  # Payment Methods endpoints
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    RegisterRequest:
      type: object
      required:
        - email
        - password
        - firstName
        - lastName
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        firstName:
          type: string
        lastName:
          type: string
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    RefreshTokenRequest:
      type: object
      required:
        - refreshToken
      properties:
        refreshToken:
          type: string
    AuthResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: object
          properties:
            accessToken:
              type: string
            refreshToken:
              type: string
            tokenType:
              type: string
            expiresIn:
              type: integer
            user:
              $ref: '#/components/schemas/User'
    TokenResponse:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string

    # User schemas
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        phoneNumber:
          type: string
        dateOfBirth:
          type: string
          format: date
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    PaymentStatus:
      type: string
      enum:
        - PENDING
        - PROCESSING
        - COMPLETED
        - FAILED
        - REFUNDED
        - CANCELLED
...
    Payment:
      type: object
      properties:
        id:
          type: integer
          format: int64
        userId:
          type: integer
          format: int64
        travelId:
          type: integer
          format: int64
        paymentMethodId:
          type: integer
          format: int64
        amount:
          type: number
        currency:
          type: string
        status:
          $ref: '#/components/schemas/PaymentStatus'
        transactionId:
          type: string
        description:
          type: string
        createdAt:
          type: string
          format: date-time
    CreatePaymentRequest:
      type: object
      required:
        - userId
        - paymentMethodId
        - amount
      properties:
        userId:
          type: integer
          format: int64
        travelId:
          type: integer
          format: int64
        paymentMethodId:
          type: integer
          format: int64
        amount:
          type: number
        currency:
          type: string
          default: EUR
        description:
          type: string
    PaymentResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          $ref: '#/components/schemas/Payment'
    PaymentListResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: array
          items:
            $ref: '#/components/schemas/Payment'
        pageInfo:
          $ref: '#/components/schemas/PageInfo'
...
    PageInfo:
      type: object
      properties:
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
        data:
          type: object
          nullable: true
        timestamp:
          type: string
          format: date-time